requirements.txt
appdirs==1.4.4
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
dj-database-url==3.0.0
Django==5.2.3
django-extensions==4.1
django-stubs==5.2.0
django-stubs-ext==5.2.0
djangorestframework==3.16.0
gunicorn==23.0.0
idna==3.10
jedi==0.19.2
Markdown==3.8
packaging==25.0
parso==0.8.4
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptpython==3.0.30
Pygments==2.19.1
requests==2.32.4
sqlparse==0.5.3
types-PyYAML==6.0.12.20250516
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
whitenoise==6.9.0
django-filter==25.1
pytz==2025.2
